Block Xr.DeviceChange event without FeaturePolicy and add tests
While both requestSession and supportsSession reject without the WebXr
Feature Policy being enabled, DeviceChange events were still firing.
This makes it so that those events do not fire unless the feature policy
is set.  Given that this is a decision based on events from the system,
and not user, if the FeaturePolicy is not set, no reporting is done.

This also adds a web platform test with the expected behavior of all of
the xr functions when the feature policy is not enabled.

Bug:1003842
Change-Id: Ie9ae16b5c1d863b730e556b511c024bae8a4503c