Split skip-not-fully-active.html into two tests
We'd like to prioritize loading and compositing tasks till FCP and this
appears to expose a problem with the test. The timing of the script on
line 6 with respect to the second promise test is a bit fragile.

If the iframe script runs before the promise then
document.querySelector('iframe') returns null and the test fails.

This timing issue seems unrelated to the test itself, so I propose
splitting the test into two as the simplest fix.

Bug: 971191
Change-Id: I65384093ac8f4616d64fc21ee692d145be1655a6